Ability to preload additional custom services
I am working on the SMART App Launcher (http://launch.smarthealthit.org/), adding support for CDS Hooks as new "launch type". In this mode people enter an URL that points to their service discovery endpoint and then click a button that launches the CDS sandbox as SMART app in new tab. By launching it as app we already have the auth handled and the `fhirBaseUrl` properly set up. The only missing piece was a way to "inject" the user's service discovery URL and to add the new services to the list. We implemented this by returning new custom property called `serviceDiscoveryURL` to the access token response. This way, when the sandbox authorizes itself, it will already know if new service will have to be loaded and load them before the initial render.

I kept the code changes to the minimum and it should not affect the CDS Sandbox in any way. However, we do introduce new proprietary property to the auth token response. I hope other parties might be interested in this as well.